corrected according to the comments. The ClaimsPrincipal builder has been created

+++ WebApi/WebApiAuthenticate/Requests/ConfirmEmailRequest.cs
@@ -7,4 +7,5 @@
 /// <param name="NewEmail">Новый Email, который подтверждают</param>
 public record ConfirmEmailRequest(
     Guid Id,
-    string NewEmail, int Code) : BaseRequest<Guid>(Id);
+    string NewEmail,
+    ushort Code) : BaseRequest<Guid>(Id);
